added filters (sort of), set some new constraints, fixed resizing issue

+++ Findr/FilterGroup.h
@@ -25,6 +25,6 @@
 
 - (void)toggleCollapsed:(int)selectedRow;
 - (Filter *)getCurrentSelection;
-- (int)displayRowsWhenCollapsed;
+- (NSInteger)displayRowsWhenCollapsed;
 
 @end
